requirements.txt
altair==4.2.0
folium==0.12.1.post1
numpy==1.22.0
pandas==1.4.0
Pillow==9.4.0
pywin32==303
streamlit==1.8.1
streamlit_folium==0.6.4
